requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
block-disposable-email==2.0.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
disposable-email-domains==0.0.117
Django==5.1.6
django-email-disposable-checker==0.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-spectacular==0.28.0
email_validator==2.2.0
environ==1.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==25.0
password-validator==1.0
phonenumbers==8.13.54
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pycryptodomex==3.21.0
PyJWT==2.10.1
pyotp==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
